Implement iterative
Create a binary tree as below to test the program.
             5
            / \
           2   1
          / \    \
         3   6   4
In this file, I continuously search the left node, move these left
nodes to right node, and the original right node is moved to the tail
of that moved left node.

In CreateNode function, it can create node by call this function and
text ‘L’, ‘LL’ or other else.