Implement recursive
Create a binary tree as below to test the program.
             5
            / \
           2   1
          / \    \
         3   6   4
In this recursive file, a binary tree is built as the general way, init
and malloc required memory, apply preorder method to search the whole
tree.

In CreateNode function, it can create node by call this function and
text ‘L’, ‘LL’ or other else.

Create output file(.txt) for gnuplot and using astyle.