Revise test method and implement sched_setscheduler
I use rand() to generate random characters for the smallese_character function,
and test this function from 'a' to 'z' in order to know whether the answer is
correct.

Use sched_setscheduler to make this program is execute for the top priority,
and get the execution time more objective.

Use astyle.