Revise the correctness of the code and test method
In previous version I use fixed pattern to test the correctness of the function,
and the space of the array is malloced wrong, and lead to the wrong answer in
some test pattern, so in this version I revise those mistakes.

I use a .txt file to save the test patterns and the answer to the pattern,
it can execute and test the function same time.

Use astyle.